robot_inspector/requirements.txt
boto3>=1.42.0
botocore>=1.42.0
pillow>=12.0.0
opencv-python>=4.8.0
pyyaml>=6.0
numpy>=2.0.0
